dhcp_client_utils: fix multiple udhcpc processes are triggered after every WAN refresh (cable plugin/plug out) ( OFW-2847 )
[Description]
Multiple udhcp client process are created when cable is plug in and out.When WAN is up in BOSA unit,wanmanager is demonizing udhcpc as child process, and getting exited.
one udhcpc in zombie state and other udhcpc in active. When plugging out fiber cable, stop_udhcpc function is trying to kill zombie process.

[Solution]
As a fix added waitpid to wait for child process to exit so that when wan is down, it can stop active udhcpc process.

[Testing Performed]
1. Provision the gateway in Dual mode.
2. Bootup the device and check for udhcpc process.There should be no udhcpc zombie process.
3. Pull out fiber cable and check ps -eaf | grep udhcpc. There should not be udhcpc process.
4. Plugin fiber cable and check ps command. Only one udhcpc process should run with appropriate arguments.

Signed-off-by: Saikiran Vanamala<svanamala@libertyglobal.com>