dhcp_client_utils: remove quotation marks from option 60 when forming udhcpc command line ( OFW-2834 )
[Issue]
The start_exe() function does not handle quotes when parsing the
command line options string so any quotes within the command line
string are passed directly to udhcpc. This results in quotation marks
being observed (via pcap) in the option 60 string.

This is a bug/limitation in start_exe() which only splits the command
line based on spaces. ie it does not parse and remove quotes or seem
to support any alternative way to pass command line options with
spaces...

The real fix is to update start_exe(), however since currently the
option 60 string is known to not include spaces simply removing the
quotes is a temporary workaround.

[Solution]
Remove quotation marks while forming the cmd line.

[Testing Performed]
Check option 60 if it as extra quotation marks or not using tcpdump

Signed-off-by: Oussema ben abdallah(obenabdallah@libertyglobal.com)