dhcp_client_utils: send option 43 in DHCPv4 discover message ( OFW-3172 )
The option 43 should be included in the option 55

BUG DESCRIPTION:
The option 43 is not included under the option 55

ROOT CAUSE:
Included option 43 is not under option55. It was send seperately

FIX DESCRIPTION:
Included option43 under option 55 by adding it to the requested list

TESTING PERFORMED:
Steps to Reproduce the Issue :
1.Reboot the DUT and start a packet capture
2.Once eRouter get an IP, stop the packet capture and check the option 55 of DHCP Discover/Request.

Upstream Candidate: No. This is part of LGI-specific DHCP requirements.

Signed-off-by: Sreeja S Kumar (srekumar.contractor@libertyglobal.com)