dhcp_client_utils: modify dhcpv4 client suboptions to send vendor information ( OFW-3265 )
use option 125 instead of option 43 by dhcpv4 client to send manufacturer/product and model informations
as per TR-124.

Change the type of option 61 from 00 to 255 as per RFC4361 instead of
RFC3315.

Sub-options of Vendor-Identifying Vendor-Specific Information Option 125 in not send in DHCPv4.

[Issue]
Option 125 was not implemented

[Description]
This OFW was in CCB. As per the instruction following changes are made. MVXREQ-3598
1. Remove send option 43. As its cable lab specific.
2. Add option 125
3. Remove 125 from the request option 55.
4. Change the type of option 61 from 00 to 255.

[Solution]
Implement changes as per MVXREQ-3598

[Testing]
1. Boot up the gateway
2. Check the bootp options are right as per the description

Upstream Candidate: No. LGI specific DHCP requirement.

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)